backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain-google-genai==2.0.5
langchain-core==0.3.28
python-dotenv==1.0.1
pydantic==2.10.3
Pillow==11.1.0
langchain-openai
watchfiles==1.1.1
langgraph==0.2.62
